Add displaying news in html or md. (#82)
It would be helpful to allow the launcher to understand and display
both md and html news items. To accomplish this, we display the raw
html of the news item if it begins and ends with opening/closing tag
characters respectively. Otherwise it defaults to displaying markdown.
If no news items, or none selected, default to empty string.